+++ script/tool/CHANGELOG.md
@@ -3,6 +3,8 @@
 - Adds a new `update-snippets` command to maintain README files using the
   `code-excerpter` package from flutter/site-shared.
 - `license-check` now ignores submodules.
+- Allows `make-deps-path-based` to skip packages it has alredy rewritten, so
+  that running multiple times won't fail after the first time.
 
 ## 0.8.2+1